Parallelize the deserialization of each FusionExecutorCache
